A new functionality has been added to the DIM server which allows for the configurable DTC <-> HW_address mapping.
In two words, a set of files like M*-*/DTCmap.dat appears in the module configuration directories
when ResetFEE command run for the first time. They are filled with the «hardcoded» maps. Later on,
one can edit these files. Let’s see this on the real example when FEE card FEE25 was plugged
into the new DTC port 35. The following changes were made in the file M3-3/DTCmap.dat:

        «25 21»  -->   «35 21»

This means that FEE card with hardware address 21 is not plugged in the DTC port 25 anymore.
Instead, it plugged in the port 35! Note that its hardware address 21 remains the same
which preserves the correct mapping between hardware address and (X,Z) position on the module plane.
Additionally, port 35 was enabled in the file FEEmask.dat which defines which FEE cards
should be switched on after the running of ResetFEE M3-3 command.